refactor(tee): deduplicate MockRegistry test helper in nitro-host
Amp-Thread-ID: https://ampcode.com/threads/T-019d759e-a470-769b-a467-a702b36dd9a5
